feat: update output to indicate what stack will be deployed
